+++ doc/CHANGELOG.rst
@@ -5,6 +5,10 @@ This changelog *only* contains changes from the *first* pypi release (0.5.4.3) o
 
 Latest Changes:
 - **1.2.1_dev0 - 2020-11-29**
+
+  - JPype scans jar files and rebuilding missing directories to allow imports
+    from stripped and obfuscated jar files.
+
 - **1.2.0 - 2020-11-29**
 
   - Added builds for Python 3.9.  Python 3.9 on Windows is currently failing